Serve on ephemeral port by default
The TCP port we bind to and serve on should ideally be not of relevance
to anybody -- we write it into the registry's config.json and that
exposes it to Cargo, but there is no need for a user to care about it
(as it stands).
As such, in order to prevent unnecessary collisions, with this change we
just use an ephemeral port altogether. This step basically removes the
need for clients to worry about which port may be free on the system.