Do not provide config path to Index::ensure_config
The Index::ensure_config method is a bit inconsistent in the sense that
it accepts an arbitrary path to what it assumed to be a config file, but
then internally it actually uses the file name config.json for error
reporting purposes.
Now, we deliberately decided to just use config.json in errors as
opposed to the full .with_context(|| format!(...)) shebang, just for the
sake of keeping the code concise. Hence, with this change we move the
creation of the path to the config.json file into the
Index::ensure_config method itself.