Try to bind to previously used port
With the change to use an ephemeral and kernel-assigned port we are now
faced with another problem: with every start of the registry we
potentially use a different port. Because this port ends up being
written into the index' config.json, which in turn is version
controlled, we may create a "useless" commit with every start of the
program.
To fix this behavior, with this change we peek into the config.json file
and try to read the previously used port. If the user asked for a kernel
assigned port, we overwrite this setting by trying to bind to the
previously used port. Only if that fails for some reason we may let the
kernel assign the port.