Make response function more general purpose
We use a small helper function, response, to craft the response to an
HTTP request we served, handling errors correctly. So far this function
could only work with empty success messages. In the future, we need it
to be more generally usable with anything that can be converted into a
response as understood by warp.
To that end, this change adjusts the function to be able to handle
non-empty success values as well.