Use warp::reply::json instead of manually doing JSON conversion
As it turns out wrap comes with a short cut for responding with an
object that can be JSON encoded. With this change we switch over to
using it, removing the need to to the "manual" conversion we used
earlier (and also eliminating the fallback we have, although presumably
the error is just swallowed now).