Prepare for serving git repository over HTTP
With an upcoming change we would like to serve the git repository
containing the index over regular HTTP -- in an attempt to make
accessing it more flexible. In addition to just using the local file
system, this is another way of publishing the index and making it
usable to Cargo.
As it turns out, in order to do that we actually need to prepare the
repository somewhat. With this change we invoke git update-server-info
in the repository to perform exactly this kind of preparation. In a
nutshell, this command will, among other things, create .git/info/refs,
which is the first file that is being transferred when repository
contents are being fetched.