Add support for serving the index over HTTP
In the past we solely relied on the index being accessible through the
file system. To make the registry easier accessible in multi-host
contexts, this change adds support for serving it over HTTP. To do
that, really all we have to do is serve the .git repository. However, we
rely on the update-server-info step to prepare the repository.

Furthermore, by simply serving the .git repository we are relying on
git's dumb transport mode [0]. Unfortunately, this transport mode is not
supported by libgit2, which is what Cargo uses by default. As such,
Cargo will not be able to download the index. In order to support this
transport mode, we need to tell Cargo to use a regular git invocation
for retrieving the index. Hence, the following configuration should be
added to Cargo's configuration (cargo.toml):

  [net]
  git-fetch-with-cli = true

[0]: https://www.git-scm.com/docs/http-protocol