Implement download handler using redirection
With this patch we change the way with which we implement the
/api/v1/crates/../../download handler. Specifically, instead of
"manually" implementing the .crate file lookup, open the file, and send
it back, we use a redirection to a new endpoint that serves the contents
of the root directory where all .crate files are stored. This way we
we make better use of warp's functionality that we already use
elsewhere.