Move index data into dedicated index/ folder
So far we stored both the index as well as the actual crate data
directly in the directory that the user provided. With this change we
instead move the index into a separate sub-directory, index/. In so
doing we align our file system layout with that of cargo-local-registry,
which means that both programs can be used to managed the same registry.