Cache build artifacts in CI
CI end-to-end test times are rather high because of the large number of
dependencies that are compiled on every run. In order to get them back
under control, with this change we cache the build artifacts from the
target/ directory as used by cargo as well as its home directory. The
cache is available not only between jobs of a single pipeline run, but
also across different pipeline runs. In our case, the cache will be
populated only by the build step. All subsequent jobs will merely read
from it. Because we have the cache available between jobs as well, we
can actually drop the usage of artifacts for transfer between stages.