Fix HTTP based retrieval of published crates
The recent change moving all index data into the index/ folder broke
HTTP based access of crates. The reason is that cargo assumes that the
Git repository that it fetches has the index at its root. Because we
violated that constraint, published crates were no longer found.
We could move the Git repository below index/ as well, but that would
mean that we would loose tracking of the actual crate tar balls, which
is a core feature that we do not want to miss (as it allows rolling back
both the meta data in the index as well as the actual crate data in
unison, and not doing will likely result in all kinds of mismatches
because of the checksumming employed).
With this change, we instead just make index/ a symbolic link to the
parent. Going down this route has the major advantage that this is not a
backwards compatibility breaking change and will not require any change
to existing registries.