Add index symlink in registry to created Git repository
With a recent change we introduced a recursive symbolic link to the
created Git repository containing the registry. However, we did not
actually commit the file itself to Git. That's arguably fine, as in the
case where the repository is cloned the registry would just recreate it.
But we can envision cases where the user may want this symlink before
starting the registry itself.
In order to cater to such cases, this change commits the symbolic link
to Git. In so doing we also make the code more similar in nature to what
we already have with the ensure_config method.