Rename my-registry to e2e-test-registry for end-to-end tests
We use the registry name my-registry in the README and so chances are
that users may actually follow this very example. If they did, they may
encounter test failures, because this setting could end up overwriting
what we use in tests.
To work around issues like this, with this change we opt for the name
e2e-test-registry for the registry used for testing instead.