Add test checking that we can instantiate Index multiple times
We do a bunch of conditional initialization as part of creating an
index. For example, we create a symbolic link within the registry
directory that is persisted on the file system.
In order to ensure that we can handle existence of such state, this
change introduces a test that instantiates the Index in the same
directory two times.