Update tracing-subscriber dependency dependency to 0.3
This change updates the tracing-subscriber dependency we consume to
version 0.3.
Note that we stop emitting local time stamps in traces with this change
(i.e., they will now be for the UTC time zone). The new version comes
with a 'local-time' feature state switch but unless we jump through huge
hoops (and have to accept unsound code), trace messages are just
silently swallowed. Please refer to
https://github.com/tokio-rs/tracing/issues/1688 for the glory details.
We could use the recently added OffsetTime, but it has the downside of
not reacting to time zone changes, which may cause confusion for a long
running service. There is nothing wrong with emitting UTC time stamps,
though, so that is what we do.