Fix "warnings" emitted when running end-to-end tests
Our end-to-end tests publish dummy crates to our dummy history.
Currently, the cargo publish command does not explicitly specify the
registry to publish to, though, and so a warning is printed:
  > Found `e2e-test-registry` as only allowed registry.
  > Publishing to it automatically.

As the message conveys, there is only one allowed registry and that is
the expected behavior.
However, given that we have no interest in seeing this warning, with
this change we explicitly specify the registry as well, in order to
silence it.