Reduce log spam by de-duplicatring request traces
Apparently wrap applies a 'with' filter attached to one of the routes to
all of them. In our case that resulted in request traces being
duplicated four times, which caused a considerable amount of log spam.
With this change we consolidate these filters into a single one,
reducing said spam in the process.