Build & test on more platforms
There is nothing conceptually preventing the program to run on platforms
other than Linux, but we never tested doing so.
With this change we add CI jobs for building and testing on MacOS and
Windows to have some assurance that the program works on these platforms
as we expect it to.