Limit Depandabot Cargo lock file updates to one PR
In the past we have hit a case where two Rust package update PRs were
merged and that resulted in a merge conflict that wasn't detected
by either change.
To prevent such cases from happening in the future, limit the number of
outstanding PRs to a single one, to serialize their creation completely.