Remove Cargo.lock
The Cargo.lock file was added only to serve the creation of the Docker
image, as it aided with reproducibility. With Docker support gone, there
is no need for us to manage this file. If users care to use specific
versions of dependencies, they are advised to manage a lock file
themselves. The only guarantee we provide is that there exist versions
of dependencies that allow a build with the stated minimum supported
Rust version.