Remove transitive dependency version hacks
In the past we required the specification of unused direct dependencies
to make sure that our minimum supported Rust version build works
properly. By now it appears as if warp has fixed their broken version
specifications and if we bump it to 0.3.7, we can drop this hack. So do
just that.